+++ src/mock/task.js
@@ -1,6 +1,7 @@
-const TASK_COUNT = 20;
-
+import {getRandomItem, getRandomBoolean, generateRepeatingDays, getRandomDate} from "../utils";
 import {COLORS} from "../const";
+
+const TASK_COUNT = 20;
 
 const DefaultRepeatingDays = {
   "mo": false,
@@ -20,41 +21,8 @@ const DescriptionItems = [
   `Пройти интенсив на соточку`,
 ];
 
-const getRandomNumber = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-const getRandomItem = (arr) => {
-  return arr[getRandomNumber(0, arr.length - 1)];
-};
-
-const getRandomBoolean = () => Math.random() > 0.5;
-
-const generateRepeatingDays = () => {
-  const randomDay = getRandomItem(PropertyDefaultRepeatingDays);
-
-  return Object.assign({}, DefaultRepeatingDays, {
-    [randomDay]: getRandomBoolean(),
-  });
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomNumber(0, 7);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-  targetDate.setHours(getRandomNumber(0, 12), getRandomNumber(0, 59));
-
-  return targetDate;
-};
-
-
 const generateMock = () => {
-  const dueDate = Math.random() > 0.5 ? null : getRandomDate();
+  const dueDate = getRandomBoolean() ? null : getRandomDate();
 
   return {
     color: getRandomItem(COLORS),
@@ -62,7 +30,7 @@ const generateMock = () => {
     isArchive: getRandomBoolean(),
     isFavorite: getRandomBoolean(),
     dueDate,
-    repeatingDays: dueDate ? DefaultRepeatingDays : generateRepeatingDays(),
+    repeatingDays: dueDate ? DefaultRepeatingDays : generateRepeatingDays(PropertyDefaultRepeatingDays),
 
   };
 };
